refactor: 패키지 기본 구조 세팅
