+++ backend/src/main/java/hanglog/category/presentation/CategoryController.java
@@ -1,0 +1,4 @@
+package hanglog.category.presentation;
+
+public class CategoryController {
+}
