+++ backend/src/main/java/hanglog/trip/dto/response/PlaceResponse.java
@@ -1,6 +1,6 @@
 package hanglog.trip.dto.response;
 
-import hanglog.category.CategoryResponse;
+import hanglog.category.dto.CategoryResponse;
 import hanglog.trip.domain.Place;
 import java.math.BigDecimal;
 import lombok.Getter;
