+++ backend/src/test/java/hanglog/trip/fixture/ExpenseFixture.java
@@ -1,6 +1,6 @@
 package hanglog.trip.fixture;
 
-import hanglog.category.Category;
+import hanglog.category.domain.Category;
 import hanglog.expense.Expense;
 
 public final class ExpenseFixture {
